% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cas_data_akl_2018_proj}
\alias{cas_data_akl_2018_proj}
\title{2018 Auckland Crash Analysis (CAS) Data}
\format{
An sf dataframe with 12,631 rows and 73 variables.
Below are the variables used in road safety analysis :
\describe{
 \items{crashSeverity}{}
 \items{crashYear}{}
 \items{minorInjuryCount}{}
 \items{seriousInjuryCount}{}
 \items{non_injury_crash}{}
 \items{geometry}{an sf point geometry type with crs 2193}

}
}
\source{
\url{https://www.nzta.govt.nz/safety/partners/crash-analysis-system/}
}
\usage{
cas_data_akl_2018_proj
}
\description{
This is a dataset containing traffic crashes that occurred in 2018 in Auckland Region
reported by New Zealand Police.
}
\keyword{datasets}
