% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_routes.R
\name{get_routes}
\alias{get_routes}
\title{Getting the routes of multiple geographic coordinate system (lat/lon) points}
\usage{
get_routes(lat_source, lng_source, lat_dest, lng_dest, weight)
}
\arguments{
\item{lat_source}{The latitude coordinates of the source point, needs to be a vector}

\item{lat_dest}{The latitude coordinates of the destination point, needs to be a vector}

\item{lng_dest}{The longitude coodinates of the destination point, needs to be a vector}

\item{lng_aource}{The longitude coodinates of the source point, needs to be a vector}
}
\description{
This function returns a dataframe of where each row an sf linesegment object from the latllon points
}
