% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{jtw_driving_akl}
\alias{jtw_driving_akl}
\title{2018 Auckland Driving Journey to Work Data}
\format{
A dataframe with 16035 rows and 24 variables.
Below are the variables used in road safety analysis :
\describe{
 \items{SA2_code_usual_residence_address}{}
 \items{SA2_name_usual_residence_address}{}
 \items{SA2_usual_residence_easting}{}
 \items{SA2_usual_residence_northing}{}
 \items{SA2_code_workplace_address}{}
 \items{SA2_name_workplace_address}{}
 \items{SA2_workplace_easting}{}
 \items{SA2_workplace_northing}{}
 \items{Drive_a_private_car_truck_or_van}{}
 \items{Drive_a_company_car_truck_or_van}{}
 \items{start_lat}{}
 \items{start_lon}{}
 \items{end_lat}{}
 \items{end_lon}{}
 \items{total_driving}{total count of drivers}


}
}
\source{
\url{https://datafinder.stats.govt.nz/table/104720-2018-census-main-means-of-travel-to-work-by-statistical-area-2/}
}
\usage{
jtw_driving_akl
}
\description{
The is a dataset containing coordinate pairs of start and end of a
drivers commute to usual residence. The dataset was extracted from Statistics
New Zealand 2018 Census main means of work.
}
\keyword{datasets}
