% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sa2_clipped_geo_akl}
\alias{sa2_clipped_geo_akl}
\title{Statistics Area 2 clipped of Auckland}
\format{
An sf dataframe with 564 rows and 7 variables:
\describe{
 \items{SA22021_V1}{}
 \items{SA22021__1}{}
 \items{SA22021__2}{}
 \items{LAND_AREA_}{}
 \items{Shape_Leng}{}
 \items{Shape_Leng}{}
 \items{geometry}{ Multipolygon geometry type with lat/lon (EPSG code 4326) crs}
}
}
\source{
\url{https://catalogue.data.govt.nz/dataset/statistical-area-2-2018-clipped-generalised}
}
\usage{
sa2_clipped_geo_akl
}
\description{
This dataset contains the clipped statistics area 2 boundaries in Auckland from
Statistics New Zealand.
}
\keyword{datasets}
