% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_route_intersects.R
\name{get_route_intersects}
\alias{get_route_intersects}
\title{Getting the polygons or route intersects}
\usage{
get_route_intersects(routes, polygon, inverse = FALSE)
}
\arguments{
\item{routes}{An sf dataframe with a LINESTRING geometry type that represents routes of multiple
source-destination points.}

\item{polygon}{An sf dataframe with a MULTIPOLYGON geometry.}

\item{inverse}{The default is FALSE. If FALSE, returns a list of routes length
and contains the indices of the polygon that intersected with the routes. If
TRUE then returns a list of polygon and contains the indices of the routes that
intersected with polygon.}
}
\value{
A list of vectors where a polygon has intersected with a given route.
Each vector in the list represents the series of polygons that has  route has
intersected and vice versa for routes.
}
\description{
This function gets the vector of the polygon ids or codes where a given route
has intersected
}
\examples{
\dontrun{

data(sa2_clipped_geo_akl)
data(jtw_akl_routes)

## initialising the router for New Zealand
ghroute::router(osm.file = "osm/new-zealand-latest.osm.pbf")

##getting the routes
jtw_akl_routes <- get_routes(as.matrix(jtw_driving_akl[c("start_lat", "start_lon", "end_lat", "end_lon")]))

get_route_intersects_test <-  get_route_intersects(jtw_akl_routes, sa2_clipped_geo_akl)

}

}
